+++ libs/cardano-ledger-conformance/src/Test/Cardano/Ledger/Conformance/Spec/Conway.hs
@@ -18,7 +18,7 @@ spec = describe "Conway conformance tests" $ do
   prop "GOV" $ conformsToImpl @"GOV" @ConwayFn @Conway
   prop "CERT" $ conformsToImpl @"CERT" @ConwayFn @Conway
   xprop "RATIFY" $ conformsToImpl @"RATIFY" @ConwayFn @Conway
-  prop "GOVCERT" $ conformsToImpl @"GOVCERT" @ConwayFn @Conway
+  xprop "GOVCERT" $ conformsToImpl @"GOVCERT" @ConwayFn @Conway
   xprop "ENACT" $ conformsToImpl @"ENACT" @ConwayFn @Conway
   prop "DELEG" $ conformsToImpl @"DELEG" @ConwayFn @Conway
   prop "POOL" $ conformsToImpl @"POOL" @ConwayFn @Conway